Rename member details component for discoverability
